Update docker install and setup

+++ .chezmoitemplates/install_ubuntu.sh
@@ -308,17 +308,24 @@ if ! command -v fastfetch &>/dev/null; then
 fi
 {{ end }}
 
-if ! command -v docker &>/dev/null; then
-  sudo apt-get update
-  sudo apt-get install -y ca-certificates curl gnupg lsb-release
-  sudo mkdir -m 0755 -p /etc/apt/keyrings
-  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
-  echo "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list >/dev/null
-  sudo apt-get update
-  sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
-  sudo groupadd docker
-  sudo usermod -aG docker $USER
-  newgrp docker
+if ! dpkg -s docker-ce &>/dev/null; then
+  if command -v snap &>/dev/null; then
+    if snap list | grep docker &>/dev/null; then
+      sudo snap remove docker
+    fi
+  fi
+  if ! command -v docker &>/dev/null; then
+    sudo apt-get update
+    sudo apt-get install -y ca-certificates curl gnupg lsb-release
+    sudo mkdir -m 0755 -p /etc/apt/keyrings
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+    echo "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list >/dev/null
+    sudo apt-get update
+    sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
+    sudo groupadd docker
+    sudo usermod -aG docker $USER
+    newgrp docker
+  fi
 fi
 
 {{ if not .isWSL }}
